+++ api/index.js
@@ -9,7 +9,7 @@ const path = require("path");
 const userRoute = require("./routes/users")
 const authRoute = require("./routes/auth")
 const postRoute = require("./routes/posts")
-
+const commentRoute = require("./routes/comments")
 
 dotenv.config();
 
@@ -54,7 +54,7 @@ app.post("/api/upload", upload.single("file"), (req, res) => {
 app.use("/api/users", userRoute);
 app.use("/api/auth", authRoute);
 app.use("/api/posts", postRoute);
-
+app.use("/api/comments", commentRoute);
 
 const port = 8000
