+++ client/src/components/post/Post.jsx
@@ -14,7 +14,7 @@ export default function Post({post}) {
 
     const [like, setLike] = useState(post.likes.length)
     const [isLiked, setIsLiked] = useState(false)
-
+    const [comme , setComm] = useState([]);
     const { user: currentUser } = useContext(AuthContext); // On donne l'alias currentUser à user, car user existe déjà plus bas (c'est l'auteur du post)
 
     const likeHandler = () => {
@@ -26,6 +26,10 @@ export default function Post({post}) {
         setLike(isLiked ? like-1 : like+1)
         setIsLiked(!isLiked)
     }
+
+    const ajoutcomm = () => {
+        const pos = document.getElementById('');
+    } 
 
     useEffect ( () => {
         setIsLiked(post.likes.includes(currentUser._id)) // Si l'array de likes comporte le currentUser -> true, sinon false.
@@ -40,6 +44,14 @@ export default function Post({post}) {
         }
         fetchUser();
     }, [post.userId])
+
+    useEffect ( () => {
+        const fetchComme = async () => {
+            const res = await axios.get(`/posts/${post._id}/comments`);
+            setComm(res.data);
+        }
+        fetchComme();
+    }, [])
 
     return (
         <div className="post">
@@ -70,7 +82,19 @@ export default function Post({post}) {
                     <div className="postBottomRight">
                         <span className="postCommentText">{post?.comments.length} comments</span>
                     </div>
-
+                </div>
+                
+                <div className="comme">
+                    <button className="commebtn">Commentaire</button>
+                    <div className="commeall">
+                        {comme.map((r)=>(
+                            <div><p>{r.content}</p> <p id="rep">{r.postId}</p> </div> 
+                        )
+                        )}
+                    </div>
+                    <label htmlFor="">Commentaire</label>
+                    <input type="text" id="comme" />
+                    <button>envoie</button>
                 </div>
             </div>
         </div>
